wip split timer and tasks into two concerns
